refactor for optional values

+++ src/main/java/com/example/learningdashboard/dtos/CategoryItemDto.java
@@ -2,21 +2,21 @@ package com.example.learningdashboard.dtos;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Setter
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 public class CategoryItemDto {
-
     private String id;
     @NotBlank
     private String type;
-
     @NotBlank
     private String color;
-
     @NotNull
-    private int upperThreshold;
+    private float upperThreshold;
 }
